@@ -10,8 +10,8 @@ import 'package:flutter_uploader_example/upload_item_view.dart';
 /// Shows the statusresponses for previous uploads.
 class ResponsesScreen extends StatefulWidget {
   ResponsesScreen({
-    Key key,
-    @required this.uploader,
+    Key? key,
+    required this.uploader,
   }) : super(key: key);
 
   final FlutterUploader uploader;
@@ -21,8 +21,8 @@ class ResponsesScreen extends StatefulWidget {
 }
 
 class _ResponsesScreenState extends State<ResponsesScreen> {
-  StreamSubscription<UploadTaskProgress> _progressSubscription;
-  StreamSubscription<UploadTaskResponse> _resultSubscription;
+  StreamSubscription<UploadTaskProgress>? _progressSubscription;
+  StreamSubscription<UploadTaskResponse>? _resultSubscription;
 
   Map<String, UploadItem> _tasks = {};
 
@@ -44,7 +44,7 @@ class _ResponsesScreenState extends State<ResponsesScreen> {
       setState(() => _tasks = tmp);
     }, onError: (ex, stacktrace) {
       print('exception: $ex');
-      print('stacktrace: $stacktrace' ?? 'no stacktrace');
+      print('stacktrace: $stacktrace');
     });
 
     _resultSubscription = widget.uploader.result.listen((result) {
@@ -54,12 +54,12 @@ class _ResponsesScreenState extends State<ResponsesScreen> {
       var tmp = <String, UploadItem>{}..addAll(_tasks);
       tmp.putIfAbsent(result.taskId, () => UploadItem(result.taskId));
       tmp[result.taskId] =
-          tmp[result.taskId].copyWith(status: result.status, response: result);
+          tmp[result.taskId]!.copyWith(status: result.status, response: result);
 
       setState(() => _tasks = tmp);
     }, onError: (ex, stacktrace) {
       print('exception: $ex');
-      print('stacktrace: $stacktrace' ?? 'no stacktrace');
+      print('stacktrace: $stacktrace');
     });
   }
 
