@@ -5,12 +5,11 @@ abstract class Upload {
   /// Default constructor which specicies a [url] and [method].
   /// Sub classes may override the method for developer convenience.
   const Upload({
-    @required this.url,
-    @required this.method,
+    required this.url,
+    required this.method,
     this.headers = const <String, String>{},
     this.tag,
-  })  : assert(url != null),
-        assert(method != null);
+  });
 
   /// Upload link
   final String url;
@@ -19,10 +18,10 @@ abstract class Upload {
   final UploadMethod method;
 
   /// HTTP headers.
-  final Map<String, String> headers;
+  final Map<String, String>? headers;
 
   /// Name of the upload request (only used on Android)
-  final String tag;
+  final String? tag;
 }
 
 /// Standard RFC 2388 multipart/form-data upload.
@@ -31,10 +30,10 @@ abstract class Upload {
 class MultipartFormDataUpload extends Upload {
   /// Default constructor which requires either files or data to be set.
   MultipartFormDataUpload({
-    @required String url,
+    required String url,
     UploadMethod method = UploadMethod.POST,
-    Map<String, String> headers,
-    String tag,
+    Map<String, String>? headers,
+    String? tag,
     this.files,
     this.data,
   })  : assert(files != null || data != null),
@@ -45,24 +44,24 @@ class MultipartFormDataUpload extends Upload {
           tag: tag,
         ) {
     // Need to specify either files or data.
-    assert(files.isNotEmpty || data.isNotEmpty);
+    assert(files!.isNotEmpty || data!.isNotEmpty);
   }
 
   /// files to be uploaded
-  final List<FileItem> files;
+  final List<FileItem>? files;
 
   /// additional data. Each entry will be sent as a form field.
-  final Map<String, String> data;
+  final Map<String, String>? data;
 }
 
 /// Also called a binary upload, this represents a upload without any form-encoding applies.
 class RawUpload extends Upload {
   /// Default constructor.
   const RawUpload({
-    @required String url,
+    required String url,
     UploadMethod method = UploadMethod.POST,
-    Map<String, String> headers,
-    String tag,
+    Map<String, String>? headers,
+    String? tag,
     this.path,
   }) : super(
           url: url,
@@ -72,5 +71,5 @@ class RawUpload extends Upload {
         );
 
   /// single file to upload
-  final String path;
+  final String? path;
 }
